tidal_auth/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

httpx==0.25.2
python-dotenv==1.0.0
rich==13.3.3
